Applied styles to the :valid :invalid :required
Added content for the required pseudo class
positioned it relative to the span that are the sibling of an input element
by using position absolute it will remove it from the normal flow
and position it relative to its parent.
now when an input field is required the user will see
a text next to it.

Added content for the valid and invalid pseudo class
positioned the generated content relative to the span
I selected the span element that are siblings of an input element
and added position relative,then selected the pseudo element of
the span via ::before or ::after and applied some style to them
if the user enters valid data a green check mark will appear
but if the user enters invalid data they will see a red x

Added style for when the user focuses or hovers over the input fields
removed the outline